+++ core/primitives/src/lib.rs
@@ -22,6 +22,7 @@ pub mod sandbox_state_patch;
 pub mod shard_layout;
 pub mod sharding;
 pub mod state;
+pub mod state_part;
 pub mod state_record;
 pub mod syncing;
 pub mod telemetry;
